Add the function `groebner_with_change_matrix`
It computes a (non-unique) standard representation of the basis
with respect to the initial generators of the ideal.

The computation mirrors the standard F4,
but also keeps track of row operations in the F4 matrix.

The standard representations are eagerly updated,
which may hurt performance a bit.

Works for AA and Nemo, not tested for DynamicPolynomials.jl