Generic coefficient type
With this commit, Groebner aims to support all
fields in AbstractAlgebra and Nemo, i.e., anything
that is a subtype of AbstractAlgebra.Field.

The implementation over AbstractAlgebra.GF(p) and
AbstractAlgebra.QQ is not affected by this commit.

For all other fields, the coefficients of
polynomials are wrapped in Groebner.CoeffGeneric.
Additionally we declare
    Groebner.CoeffGeneric <: Groebner.Coeff
so that F4 supports Groebner.CoeffGeneric.

Low-level functions now may accept and output
Groebner.CoeffGeneric.

Some more tests and benchmarks will follow.